Minesweeper
It looks like the last major function missing is having the adjecent buttons depress when both mouse buttons are depressed. It does that right now, but you have to move the mouse after holding both buttons down.

Also, flags get placed where the mouse was pressed, not where it is released.